Use only the path portion of the url to lookup mime type
Lookups were failing when query params existed on the path, resulting
in a content-type of application/octet-stream.

Using the path instead resolves this issue.

Note: in some cases, path is undefined.  This caused one of the connect
unit tests to fail: `should create a connect server`

Since connect and express seem to behave differently in some cases
perhaps it would be worthwhile updating the unit tests to run all
against each type of supported server.

related to #7
related to #10